requirements.txt
selenium
webdriver-manager
requests
beautifulsoup4
Pillow
